[Php84] [TypeDeclaration] Extract duplicated rule logic into shared services
