+++ rules/Php84/Rector/Foreach_/ForeachToArrayFindRector.php
@@ -5,18 +5,8 @@ declare(strict_types=1);
 namespace Rector\Php84\Rector\Foreach_;
 
 use PhpParser\Node;
-use PhpParser\Node\Expr\ArrowFunction;
-use PhpParser\Node\Expr\Assign;
-use PhpParser\Node\Expr\Variable;
-use PhpParser\Node\Param;
-use PhpParser\Node\Stmt\Break_;
-use PhpParser\Node\Stmt\Expression;
-use PhpParser\Node\Stmt\Foreach_;
-use PhpParser\Node\Stmt\If_;
-use Rector\NodeManipulator\StmtsManipulator;
-use Rector\Php84\NodeAnalyzer\ForeachKeyUsedInConditionalAnalyzer;
+use Rector\Php84\NodeFactory\ForeachToArrayFindFactory;
 use Rector\PhpParser\Enum\NodeGroup;
-use Rector\PhpParser\Node\Value\ValueResolver;
 use Rector\Rector\AbstractRector;
 use Rector\ValueObject\PhpVersionFeature;
 use Rector\ValueObject\PolyfillPackage;
@@ -31,9 +21,7 @@ use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 final class ForeachToArrayFindRector extends AbstractRector implements MinPhpVersionInterface, RelatedPolyfillInterface
 {
     public function __construct(
-        private readonly ValueResolver $valueResolver,
-        private readonly StmtsManipulator $stmtsManipulator,
-        private readonly ForeachKeyUsedInConditionalAnalyzer $foreachKeyUsedInConditionalAnalyzer
+        private readonly ForeachToArrayFindFactory $foreachToArrayFindFactory
     ) {
     }
 
@@ -74,87 +62,7 @@ CODE_SAMPLE
      */
     public function refactor(Node $node): ?Node
     {
-        if ($node->stmts === null) {
-            return null;
-        }
-
-        foreach ($node->stmts as $key => $stmt) {
-            if (! $stmt instanceof Foreach_) {
-                continue;
-            }
-
-            $prevStmt = $node->stmts[$key - 1] ?? null;
-            if (! $prevStmt instanceof Expression) {
-                continue;
-            }
-
-            if (! $prevStmt->expr instanceof Assign) {
-                continue;
-            }
-
-            $foreach = $stmt;
-            $prevAssign = $prevStmt->expr;
-
-            if (! $this->valueResolver->isNull($prevAssign->expr)) {
-                continue;
-            }
-
-            if (! $prevAssign->var instanceof Variable) {
-                continue;
-            }
-
-            $assignedVariable = $prevAssign->var;
-
-            if (! $this->isValidForeachStructure($foreach, $assignedVariable)) {
-                continue;
-            }
-
-            if ($this->stmtsManipulator->isVariableUsedInNextStmt(
-                $node,
-                $key + 1,
-                (string) $this->getName($foreach->valueVar)
-            )) {
-                continue;
-            }
-
-            /** @var If_ $firstNodeInsideForeach */
-            $firstNodeInsideForeach = $foreach->stmts[0];
-
-            $condition = $firstNodeInsideForeach->cond;
-            $valueParam = $foreach->valueVar;
-
-            if (! $valueParam instanceof Variable) {
-                continue;
-            }
-
-            $params = [new Param($valueParam)];
-
-            if ($foreach->keyVar instanceof Variable && $this->foreachKeyUsedInConditionalAnalyzer->isUsed(
-                $foreach->keyVar,
-                $condition
-            )) {
-                $params[] = new Param(new Variable((string) $this->getName($foreach->keyVar)));
-            }
-
-            $arrowFunction = new ArrowFunction([
-                'params' => $params,
-                'expr' => $condition,
-            ]);
-
-            $funcCall = $this->nodeFactory->createFuncCall('array_find', [$foreach->expr, $arrowFunction]);
-
-            $newAssign = new Assign($assignedVariable, $funcCall);
-            $newExpression = new Expression($newAssign);
-
-            unset($node->stmts[$key - 1]);
-            $node->stmts[$key] = $newExpression;
-
-            $node->stmts = array_values($node->stmts);
-
-            return $node;
-        }
-
-        return null;
+        return $this->foreachToArrayFindFactory->createArrayFindAssign($node, 'array_find', false);
     }
 
     public function provideMinPhpVersion(): int
@@ -165,49 +73,5 @@ CODE_SAMPLE
     public function providePolyfillPackage(): string
     {
         return PolyfillPackage::PHP_84;
-    }
-
-    private function isValidForeachStructure(Foreach_ $foreach, Variable $assignedVariable): bool
-    {
-        if (count($foreach->stmts) !== 1) {
-            return false;
-        }
-
-        $firstStmt = $foreach->stmts[0];
-        if (
-            ! $firstStmt instanceof If_ ||
-            count($firstStmt->stmts) !== 2
-        ) {
-            return false;
-        }
-
-        $assignmentStmt = $firstStmt->stmts[0];
-        $breakStmt = $firstStmt->stmts[1];
-
-        if (
-            ! $assignmentStmt instanceof Expression ||
-            ! $assignmentStmt->expr instanceof Assign ||
-            ! $breakStmt instanceof Break_
-        ) {
-            return false;
-        }
-
-        $assignment = $assignmentStmt->expr;
-
-        if (! $this->nodeComparator->areNodesEqual($assignment->var, $assignedVariable)) {
-            return false;
-        }
-
-        if (! $this->nodeComparator->areNodesEqual($assignment->expr, $foreach->valueVar)) {
-            return false;
-        }
-
-        if (! $foreach->valueVar instanceof Variable) {
-            return false;
-        }
-
-        $type = $this->nodeTypeResolver->getNativeType($foreach->expr);
-        return $type->isArray()
-            ->yes();
     }
 }
