+++ rules/TypeDeclaration/Rector/Class_/TypedPropertyFromContainerGetSetUpRector.php
@@ -6,27 +6,13 @@ namespace Rector\TypeDeclaration\Rector\Class_;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr;
-use PhpParser\Node\Expr\Assign;
 use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Expr\PropertyFetch;
 use PhpParser\Node\Expr\StaticCall;
 use PhpParser\Node\Expr\Variable;
-use PhpParser\Node\Name;
 use PhpParser\Node\Stmt\Class_;
-use PhpParser\Node\Stmt\ClassMethod;
-use PhpParser\Node\Stmt\Property;
-use PHPStan\PhpDocParser\Ast\PhpDoc\VarTagValueNode;
-use PHPStan\Reflection\ReflectionProvider;
-use PHPStan\Type\ObjectType;
-use Rector\BetterPhpDocParser\PhpDocInfo\PhpDocInfo;
-use Rector\BetterPhpDocParser\PhpDocInfo\PhpDocInfoFactory;
-use Rector\Comments\NodeDocBlock\DocBlockUpdater;
-use Rector\PhpParser\Node\BetterNodeFinder;
-use Rector\PHPStanStaticTypeMapper\Enum\TypeKind;
-use Rector\PHPUnit\NodeAnalyzer\TestsNodeAnalyzer;
 use Rector\Rector\AbstractRector;
-use Rector\StaticTypeMapper\StaticTypeMapper;
-use Rector\ValueObject\MethodName;
+use Rector\TypeDeclaration\NodeAnalyzer\SetUpAssignedPropertyTyper;
 use Rector\ValueObject\PhpVersionFeature;
 use Rector\VersionBonding\Contract\MinPhpVersionInterface;
 use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
@@ -38,12 +24,7 @@ use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 final class TypedPropertyFromContainerGetSetUpRector extends AbstractRector implements MinPhpVersionInterface
 {
     public function __construct(
-        private readonly TestsNodeAnalyzer $testsNodeAnalyzer,
-        private readonly PhpDocInfoFactory $phpDocInfoFactory,
-        private readonly StaticTypeMapper $staticTypeMapper,
-        private readonly DocBlockUpdater $docBlockUpdater,
-        private readonly BetterNodeFinder $betterNodeFinder,
-        private readonly ReflectionProvider $reflectionProvider
+        private readonly SetUpAssignedPropertyTyper $setUpAssignedPropertyTyper
     ) {
     }
 
@@ -101,104 +82,15 @@ CODE_SAMPLE
      */
     public function refactor(Node $node): ?Node
     {
-        if (! $this->testsNodeAnalyzer->isInTestClass($node)) {
-            return null;
-        }
-
-        $setUpClassMethod = $node->getMethod(MethodName::SET_UP);
-        if (! $setUpClassMethod instanceof ClassMethod) {
-            return null;
-        }
-
-        $hasChanged = false;
-
-        foreach ($node->getProperties() as $property) {
-            // type is already set
-            if ($property->type instanceof Node) {
-                continue;
-            }
-
-            if (! $property->isPrivate()) {
-                continue;
-            }
-
-            if ($property->isStatic()) {
-                continue;
-            }
-
-            // exactly one property
-            if (count($property->props) !== 1) {
-                continue;
-            }
-
-            $propertyName = $this->getName($property->props[0]);
-            if (! $this->isAssignedViaContainerGetInSetUp($setUpClassMethod, $propertyName)) {
-                continue;
-            }
-
-            $propertyPhpDocInfo = $this->phpDocInfoFactory->createFromNode($property);
-            if (! $propertyPhpDocInfo instanceof PhpDocInfo) {
-                continue;
-            }
-
-            $varType = $propertyPhpDocInfo->getVarType();
-            if (! $varType instanceof ObjectType) {
-                continue;
-            }
-
-            $propertyTypeNode = $this->staticTypeMapper->mapPHPStanTypeToPhpParserNode($varType, TypeKind::PROPERTY);
-            if (! $propertyTypeNode instanceof Name) {
-                continue;
-            }
-
-            // must be an existing object type
-            if (! $this->reflectionProvider->hasClass($propertyTypeNode->toString())) {
-                continue;
-            }
-
-            $property->type = $propertyTypeNode;
-            $this->removeVarTag($propertyPhpDocInfo, $property);
-
-            $hasChanged = true;
-        }
-
-        if ($hasChanged) {
-            return $node;
-        }
-
-        return null;
+        return $this->setUpAssignedPropertyTyper->refactorClass(
+            $node,
+            fn (Expr $expr): bool => $this->isContainerGetCall($expr)
+        );
     }
 
     public function provideMinPhpVersion(): int
     {
         return PhpVersionFeature::TYPED_PROPERTIES;
-    }
-
-    private function isAssignedViaContainerGetInSetUp(ClassMethod $setUpClassMethod, string $propertyName): bool
-    {
-        /** @var Assign[] $assigns */
-        $assigns = $this->betterNodeFinder->findInstanceOf($setUpClassMethod, Assign::class);
-
-        foreach ($assigns as $assign) {
-            if (! $assign->var instanceof PropertyFetch) {
-                continue;
-            }
-
-            $propertyFetch = $assign->var;
-            if (! $this->isName($propertyFetch->var, 'this')) {
-                continue;
-            }
-
-            if (! $this->isName($propertyFetch, $propertyName)) {
-                continue;
-            }
-
-            if ($this->isContainerGetCall($assign->expr)) {
-                return true;
-            }
-        }
-
-        return false;
     }
 
     private function isContainerGetCall(Expr $expr): bool
@@ -225,16 +117,5 @@ CODE_SAMPLE
 
         // $this->get(...)
         return $caller instanceof Variable && $this->isName($caller, 'this');
-    }
-
-    private function removeVarTag(PhpDocInfo $propertyPhpDocInfo, Property $property): void
-    {
-        $varTagValueNode = $propertyPhpDocInfo->getVarTagValueNode();
-        if (! $varTagValueNode instanceof VarTagValueNode) {
-            return;
-        }
-
-        $propertyPhpDocInfo->removeByType(VarTagValueNode::class);
-        $this->docBlockUpdater->updateRefactoredNodeWithPhpDocInfo($property);
     }
 }
